Moved v1.1.0 to legacy file
